UI Implementation
Currently, from the main menu, the user is able to Exit the application, navigate the menu to the create room screen, along with the size selections. Only 1 size is currently available until more features and fixes gets implemented. Picking the room size leads to the Lobby which allows the user to start up the Room. The user is also able to return to any previous screen if needed. Avatar selection is not ready.